requirements.txt
httpx
apscheduler
telethon
pysocks
